- Refactored the knxnetip protocol to have the ETS5 parts in a separate mspec file - Moded the ETS5 code into a shared module - Renamed the generated knxnetip and s7 driver modules to have the "passive" at the end
