+++ protocols/knxnetip/src/main/resources/META-INF/services/org.apache.plc4x.plugins.codegenerator.protocol.Protocol
@@ -17,3 +17,4 @@
 # under the License.
 #
 org.apache.plc4x.protocol.knxnetip.KnxNetIpProtocol
+org.apache.plc4x.protocol.knxnetip.Ets5Types
