+++ sandbox/pom.xml
@@ -39,11 +39,15 @@
     <module>plc-simulator</module>
 
     <module>test-java-bacnetip-driver</module>
-    <module>test-java-knxnetip-driver</module>
-    <module>test-java-passive-knxnetip-driver</module>
-    <module>test-java-s7-driver</module>
-    <module>test-java-passive-s7-driver</module>
+
     <module>test-java-df1-driver</module>
+
+    <module>test-java-knxnetip-driver</module>
+    <module>test-java-knxnetip-driver-passive</module>
+    <module>test-java-knxnetip-shared</module>
+
+    <module>test-java-s7-driver</module>
+    <module>test-java-s7-driver-passive</module>
 
     <module>test-streampipes-plc4x-adapters</module>
     <module>test-streampipes-plc4x-processors</module>
