+++ sandbox/test-streampipes-plc4x-processors/pom.xml
@@ -51,6 +51,11 @@
       <artifactId>test-java-knxnetip-driver</artifactId>
       <version>0.6.0-SNAPSHOT</version>
     </dependency>
+    <dependency>
+      <groupId>org.apache.plc4x.sandbox</groupId>
+      <artifactId>test-java-knxnetip-shared</artifactId>
+      <version>0.6.0-SNAPSHOT</version>
+    </dependency>
 
     <dependency>
       <groupId>org.streampipes</groupId>
@@ -146,11 +151,6 @@
       <groupId>org.apache.plc4x.sandbox</groupId>
       <artifactId>test-streampipes-plc4x-shared</artifactId>
       <version>0.6.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-compress</artifactId>
-      <version>1.19</version>
     </dependency>
     <dependency>
       <groupId>javax.validation</groupId>
